web-app/ai_models/requirements.txt
opencv-python
numpy
torch
ultralytics
Pillow
Flask
transformers
sentence-transformers
easyocr
requests
googlemaps
pymongo
pdf2image
poppler-utils
tqdm
langchain
llama-cpp-python